add description in README remove server port in properties add annotation @RequestMapping in ControllerCompany add annotation @Repository in RepositoryCompany change data type from Integer to Long

+++ src/main/java/ru/ageev/companyregister/service/ServiceCompany.java
@@ -8,11 +8,11 @@ public interface ServiceCompany {
 
     void create(Company company);
 
-    Company read(int id);
+    Company read(long id);
 
     List<Company> readAll();
 
-    boolean update (Company company,int id);
+    boolean update (Company company,long id);
 
-    boolean delete(int id);
+    boolean delete(long id);
 }
